% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-RiemGrassmann.R
\docType{package}
\name{package-RiemGrassmann}
\alias{package-RiemGrassmann}
\title{Inference, Learning, and Optimization on Grassmann Manifold}
\description{
Grassmann manifold \eqn{Gr(k,n)} is the set of k-planes, or \eqn{k}-dimensional subspaces in \eqn{R^n}, 
which is indeed a compact Riemannian manifold. In this package, we use a convention that each element 
in \eqn{Gr(k,n)} is represented as an orthonormal basis (ONB) \eqn{X \in \mathbf{R}^{n\times k}} where
\deqn{X^\top X = I_k}.
}
